// packages/muya/src/state/types.ts
export interface IBlockQuoteState {
    name: 'block-quote';
    meta?: {
        admonitionType?: TAdmonitionType;
    };
    children: TState[];
}
// ...
export type TLeafState
    = | IParagraphState
        | IAtxHeadingState
        | ISetextHeadingState
        | IThematicBreakState
        | ICodeBlockState
        | IHtmlBlockState
        | ILinkReferenceDefinitionState
        | IMathBlockState
        | IFrontmatterState
        | IDiagramState
        | ITableCellState;

export type TContainerState
    = | IBlockQuoteState
        | IOrderListState
        | IBulletListState
        | ITableState
        | ITaskListState
        | ITaskListItemState
        | IListItemState
        | ITableRowState
        | IFootnoteBlockState;

export type TState = TLeafState | TContainerState;